#18 [refactor] base 생성
