+++ app/src/main/java/com/sopt/now/config/BaseActivity.kt
@@ -1,0 +1,21 @@
+package com.sopt.now.config
+
+import android.os.Bundle
+import android.os.PersistableBundle
+import android.view.LayoutInflater
+import androidx.appcompat.app.AppCompatActivity
+import androidx.viewbinding.ViewBinding
+
+abstract class BaseActivity<B : ViewBinding>(private val inflate: (LayoutInflater) -> B) :
+    AppCompatActivity() {
+        protected lateinit var binding: B
+
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = inflate(layoutInflater).apply {
+            setContentView(root)
+        }
+    }
+}
