+++ app/src/main/java/com/sopt/now/config/BaseFragment.kt
@@ -1,0 +1,31 @@
+package com.sopt.now.config
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.LayoutRes
+import androidx.fragment.app.Fragment
+import androidx.viewbinding.ViewBinding
+
+abstract class BaseFragment<B : ViewBinding>(
+    private val bind: (View) -> B,
+    @LayoutRes layoutResId: Int
+) : Fragment(layoutResId) {
+    private var _binding: B? = null
+    protected val binding get() = requireNotNull(_binding)
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = super.onCreateView(inflater, container, savedInstanceState)?.let { bind(it) }
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        _binding = null
+        super.onDestroyView()
+    }
+}
