+++ app/src/main/java/com/sopt/now/data/AuthRepositoryImpl.kt
@@ -4,7 +4,7 @@ import com.sopt.now.datasource.AuthService
 import com.sopt.now.model.login.RequestLoginDto
 import com.sopt.now.model.signup.RequestSignUpDto
 import com.sopt.now.repository.AuthRepository
-import com.sopt.now.utils.BaseResponse
+import com.sopt.now.config.BaseResponse
 import retrofit2.Response
 
 class AuthRepositoryImpl(
