+++ app/src/main/java/com/sopt/now/data/MyPageRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.sopt.now.data
 import com.sopt.now.datasource.InfoService
 import com.sopt.now.model.info.UserInfo
 import com.sopt.now.repository.MyPageRepository
-import com.sopt.now.utils.BaseResponse
+import com.sopt.now.config.BaseResponse
 import retrofit2.Response
 
 class MyPageRepositoryImpl(
