+++ app/src/main/java/com/sopt/now/presentation/main/home/HomeFragment.kt
@@ -6,21 +6,12 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
+import com.sopt.now.R
+import com.sopt.now.config.BaseFragment
 import com.sopt.now.databinding.FragmentHomeBinding
 
-class HomeFragment : Fragment() {
-    private var _binding: FragmentHomeBinding? = null
-    private val binding get() = requireNotNull(_binding)
+class HomeFragment : BaseFragment<FragmentHomeBinding>(FragmentHomeBinding::bind, R.layout.fragment_home) {
     private val viewModel by viewModels<HomeViewModel>()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
-        return binding.root
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -32,10 +23,5 @@ class HomeFragment : Fragment() {
         val friendAdapter = ProfileAdapter()
         binding.rvFriend.adapter = friendAdapter
         friendAdapter.setFriendList(viewModel.mockFriendList)
-    }
-
-    override fun onDestroyView() {
-        _binding = null
-        super.onDestroyView()
     }
 }
