+++ app/src/main/java/com/sopt/now/presentation/main/mypage/MyPageFragment.kt
@@ -8,6 +8,8 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
+import com.sopt.now.R
+import com.sopt.now.config.BaseFragment
 import com.sopt.now.databinding.FragmentMypageBinding
 import com.sopt.now.model.info.UserInfo
 import com.sopt.now.utils.UiState
@@ -17,19 +19,8 @@ import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 
 @AndroidEntryPoint
-class MyPageFragment : Fragment() {
-    private var _binding: FragmentMypageBinding? = null
-    private val binding get() = requireNotNull(_binding)
+class MyPageFragment : BaseFragment<FragmentMypageBinding>(FragmentMypageBinding::bind, R.layout.fragment_mypage) {
     private val myPageViewModel: MyPageViewModel by viewModels()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentMypageBinding.inflate(inflater, container, false)
-        return binding.root
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -40,11 +31,6 @@ class MyPageFragment : Fragment() {
 
     private fun getUserInfo() {
         myPageViewModel.getUserInfo()
-    }
-
-    override fun onDestroyView() {
-        _binding = null
-        super.onDestroyView()
     }
 
     private fun observeUserInfo() {
