+++ app/src/main/java/com/sopt/now/repository/AuthRepository.kt
@@ -2,7 +2,7 @@ package com.sopt.now.repository
 
 import com.sopt.now.model.login.RequestLoginDto
 import com.sopt.now.model.signup.RequestSignUpDto
-import com.sopt.now.utils.BaseResponse
+import com.sopt.now.config.BaseResponse
 import retrofit2.Response
 
 interface AuthRepository {
