+++ app/src/main/java/com/sopt/now/repository/MyPageRepository.kt
@@ -1,7 +1,7 @@
 package com.sopt.now.repository
 
 import com.sopt.now.model.info.UserInfo
-import com.sopt.now.utils.BaseResponse
+import com.sopt.now.config.BaseResponse
 import retrofit2.Response
 
 interface MyPageRepository {
